Move Commands::Init into its own file
