+++ lib/munge/cli.rb
@@ -1,6 +1,7 @@
 require "rainbow"
 require "thor"
 
+require "munge/cli/commands/init"
 require "munge/cli/commands/build"
 require "munge/cli/commands/view"
 require "munge/cli/commands/server"
