+++ lib/munge/cli/dispatch.rb
@@ -1,20 +1,9 @@
 module Munge
   module Cli
     class Dispatch < Thor
-      include Thor::Actions
-
-      def self.source_root
-        File.expand_path("../../../../seeds", __FILE__)
-      end
-
       desc "init PATH", "Create new site at PATH"
       def init(path)
-        directory ".", path
-
-        inside(path) do
-          run_bundle("install")
-          run_bundle("binstub munge")
-        end
+        Commands::Init.new(path).call
       end
 
       desc "build", "Build site"
@@ -58,18 +47,6 @@ module Munge
 
       def symbolized_options
         Munge::Util::SymbolHash.deep_convert(options)
-      end
-
-      def run_bundle(command)
-        if Gem::Specification.find_all_by_name("bundler").any?
-          say_status :run, "bundle #{command}"
-
-          require "bundler"
-
-          ::Bundler.with_clean_env do
-            system("bundle #{command}")
-          end
-        end
       end
     end
   end
